requirements.txt
slack_bolt
requests
python-dotenv
loguru
sqlalchemy
psycopg2-binary
alembic